+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Changes ###
 
   * Fix an error in the Objective Tracker skin
-  * Reenable Rune Tacker on the RBReminder
-  * Readded Overall Score on the LFG Group Info
+  * Bumb TOC Version",
+  * Added a compatibility check for Windtools",
+  * Added two new hearthstones to LocBar & MicroBar",
+  * Added gradient text to the Tooltip Name",
+  * Updated Details Profile",